Set op new functions module for combination functions
