+++ py-polars/polars/functions/__init__.py
@@ -1,3 +1,13 @@
+from polars.functions.as_datatype import (
+    concat_list,
+    concat_str,
+    duration,
+    format,
+    struct,
+)
+from polars.functions.as_datatype import date_ as date
+from polars.functions.as_datatype import datetime_ as datetime
+from polars.functions.as_datatype import time_ as time
 from polars.functions.eager import align_frames, concat
 from polars.functions.lazy import (
     all,
@@ -10,20 +20,16 @@ from polars.functions.lazy import (
     coalesce,
     col,
     collect_all,
-    concat_list,
-    concat_str,
     corr,
     count,
     cov,
     cumfold,
     cumreduce,
     cumsum,
-    duration,
     element,
     exclude,
     first,
     fold,
-    format,
     from_epoch,
     groups,
     head,
@@ -42,14 +48,10 @@ from polars.functions.lazy import (
     rolling_cov,
     select,
     std,
-    struct,
     sum,
     tail,
     var,
 )
-from polars.functions.lazy import date_ as date
-from polars.functions.lazy import datetime_ as datetime
-from polars.functions.lazy import time_ as time
 from polars.functions.range import arange, date_range, time_range
 from polars.functions.repeat import ones, repeat, zeros
 from polars.functions.whenthen import when
